Reorganize story and fix entity delete bug
