+++ packages/databyss-editor/stories/fixtures/refEntities.js
@@ -1,0 +1,18 @@
+export const sourceFixture = {
+  _id: '5e3b2000fda293001813b1d6',
+  text: {
+    textValue:
+      'Stamenov, Language Structure, Discourse and the Access to Consciousness',
+    ranges: [],
+  },
+  citations: [{ textValue: '', ranges: [] }],
+  authors: [{ firstName: '', lastName: '' }],
+}
+
+export const topicFixture = {
+  _id: '5e3b1bc48fb28680fe26437d',
+  text: {
+    textValue: 'topic',
+    ranges: [],
+  },
+}
